backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.3
GitPython==3.1.43
pydantic==2.8.2
python-dotenv==1.0.1
google-generativeai==0.7.2
pathspec==0.12.1
openapi-spec-validator==0.7.1
